requirements.txt
mcp-sdk==1.0.0
psutil==5.8.0
pillow==8.4.0
pypdf2==1.26.0
python-dotenv==0.19.0

